+++ examples/c_tebd.py
@@ -17,7 +17,8 @@ def example_TEBD_gs_finite(L, g):
     print("finite TEBD, imaginary time evolution, L={L:d}, g={g:.2f}".format(L=L, g=g))
     model_params = dict(L=L, J=1., g=g, bc_MPS='finite', conserve=None, verbose=0)
     M = TFIChain(model_params)
-    psi = MPS.from_product_state(M.lat.mps_sites(), [0] * L, bc='finite')
+    product_state = ["up"] * model_params['L']
+    psi = MPS.from_product_state(M.lat.mps_sites(), product_state, bc=model_params['bc_MPS'])
     tebd_params = {
         'order': 2,
         'delta_tau_list': [0.1, 0.01, 0.001, 1.e-4, 1.e-5],
@@ -45,7 +46,8 @@ def example_TEBD_gs_infinite(g):
     print("infinite TEBD, imaginary time evolution, g={g:.2f}".format(g=g))
     model_params = dict(L=2, J=1., g=g, bc_MPS='infinite', conserve=None, verbose=0)
     M = TFIChain(model_params)
-    psi = MPS.from_product_state(M.lat.mps_sites(), [0] * 2, bc='infinite')
+    product_state = ["up"] * model_params['L']
+    psi = MPS.from_product_state(M.lat.mps_sites(), product_state, bc=model_params['bc_MPS'])
     tebd_params = {
         'order': 2,
         'delta_tau_list': [0.1, 0.01, 0.001, 1.e-4, 1.e-5],